server: [storage] enable restoring from corrupted installations
In the previous commit I introduced a way to verify whether the
target installation to which we are restoring data is compatible
with what we are restoring. However, this can be an issue if the
current install is corrupted.

This is why I added the `--skip-compatibility-check` flag which
tells the restore system to not attempt to verify metadata as it
might possibly be corrupted.